+++ src/reuse/header.py
@@ -54,6 +54,7 @@ from ._util import (
     detect_line_endings,
     extract_spdx_info,
     make_copyright_line,
+    merge_copyright_lines,
     spdx_identifier,
 )
 from .project import Project
@@ -127,6 +128,7 @@ def _create_new_header(
     return result
 
 
+# pylint: disable=too-many-arguments
 def create_header(
     spdx_info: SpdxInfo,
     header: str = None,
@@ -134,6 +136,7 @@ def create_header(
     template_is_commented: bool = False,
     style: CommentStyle = None,
     force_multi: bool = False,
+    merge_copyrights: bool = False,
 ) -> str:
     """Create a header containing *spdx_info*. *header* is an optional argument
     containing a header which should be modified to include *spdx_info*. If
@@ -160,10 +163,19 @@ def create_header(
                 "existing header contains an erroneous SPDX expression"
             ) from err
 
+        if merge_copyrights:
+            spdx_copyrights = merge_copyright_lines(
+                spdx_info.copyright_lines.union(existing_spdx.copyright_lines),
+            )
+        else:
+            spdx_copyrights = spdx_info.copyright_lines.union(
+                existing_spdx.copyright_lines
+            )
+
         # TODO: This behaviour does not match the docstring.
         spdx_info = SpdxInfo(
             spdx_info.spdx_expressions.union(existing_spdx.spdx_expressions),
-            spdx_info.copyright_lines.union(existing_spdx.copyright_lines),
+            spdx_copyrights,
         )
 
     new_header += _create_new_header(
@@ -232,6 +244,7 @@ def _extract_shebang(prefix: str, text: str) -> Tuple[str, str]:
     return (shebang, text)
 
 
+# pylint: disable=too-many-arguments
 def find_and_replace_header(
     text: str,
     spdx_info: SpdxInfo,
@@ -239,6 +252,7 @@ def find_and_replace_header(
     template_is_commented: bool = False,
     style: CommentStyle = None,
     force_multi: bool = False,
+    merge_copyrights: bool = False,
 ) -> str:
     """Find the first SPDX comment block in *text*. That comment block is
     replaced by a new comment block containing *spdx_info*. It is formatted as
@@ -300,6 +314,7 @@ def find_and_replace_header(
         template_is_commented=template_is_commented,
         style=style,
         force_multi=force_multi,
+        merge_copyrights=merge_copyrights,
     )
 
     new_text = f"{header}\n"
@@ -411,6 +426,7 @@ def _add_header_to_file(
     style: Optional[str],
     force_multi: bool = False,
     skip_existing: bool = False,
+    merge_copyrights: bool = False,
     out=sys.stdout,
 ) -> int:
     """Helper function."""
@@ -452,6 +468,7 @@ def _add_header_to_file(
             template_is_commented=template_is_commented,
             style=style,
             force_multi=force_multi,
+            merge_copyrights=merge_copyrights,
         )
     except CommentCreateError:
         out.write(
@@ -508,7 +525,7 @@ def add_arguments(parser) -> None:
     parser.add_argument(
         "--year",
         "-y",
-        action="store",
+        action="append",
         type=str,
         help=_("year of copyright statement, optional"),
     )
@@ -537,6 +554,11 @@ def add_arguments(parser) -> None:
         "--exclude-year",
         action="store_true",
         help=_("do not include year in statement"),
+    )
+    parser.add_argument(
+        "--merge-copyrights",
+        action="store_true",
+        help=_("merge copyright lines if copyright statements are identical"),
     )
     parser.add_argument(
         "--single-line",
@@ -659,8 +681,10 @@ def run(args, project: Project, out=sys.stdout) -> int:
 
     year = None
     if not args.exclude_year:
-        if args.year:
-            year = args.year
+        if args.year and len(args.year) > 1:
+            year = f"{min(args.year)} - {max(args.year)}"
+        elif args.year:
+            year = args.year.pop()
         else:
             year = datetime.date.today().year
 
@@ -701,6 +725,7 @@ def run(args, project: Project, out=sys.stdout) -> int:
             style=args.style,
             force_multi=args.multi_line,
             skip_existing=args.skip_existing,
+            merge_copyrights=args.merge_copyrights,
             out=out,
         )